[profiling] move profiling to new package.
