+++ pkg/config/config.go
@@ -128,11 +128,11 @@ func init() {
 	// Configure Datadog global configuration
 	Datadog = NewConfig("datadog", "DD", strings.NewReplacer(".", "_"))
 	// Configuration defaults
-	initConfig(Datadog)
+	InitConfig(Datadog)
 }
 
-// initConfig initializes the config defaults on a config
-func initConfig(config Config) {
+// InitConfig initializes the config defaults on a config
+func InitConfig(config Config) {
 	// Agent
 	// Don't set a default on 'site' to allow detecting with viper whether it's set in config
 	config.BindEnv("site")   //nolint:errcheck
