+++ pkg/config/mock.go
@@ -22,6 +22,6 @@ func Mock() *MockConfig {
 	// Configure Datadog global configuration
 	Datadog = NewConfig("datadog", "DD", strings.NewReplacer(".", "_"))
 	// Configuration defaults
-	initConfig(Datadog)
+	InitConfig(Datadog)
 	return &MockConfig{Datadog}
 }
